se trabajo y termino las validaciones solicitas

+++ src/controllers/controlador.js
@@ -16,9 +16,12 @@ const obtenerTareas = async (req, res) => {
 
 // //Obtener por id
 const obtenerTareasPorId = async (req, res) => {
+    const id = req.params.id
+    if (!id) {
+        return res.status(400).json({ error: 'El ID es requerido' });
+    }
     try {
         const connection = await dataBase()
-        const id = req.params.id
         const results = await connection.query("SELECT * FROM tasks WHERE id=?", id)
         res.status(200).json(results[0])
 
@@ -32,8 +35,17 @@ const anadirTareas = async (req, res) => {
     try {
         const connection = await dataBase()
         const { title, description, isComplete } = req.body
+        if (title.trim() === '' || title.length >= 225) {
+            return res.status(400).json({ error: 'El titulo no puede estar vacio y debe ser menor a 255 caracteres' });
+        }
+        if (description.trim() === '') {
+            return res.status(400).json({ error: 'La descripcion no debe estar vacia' });
+        }
+        if (isComplete != false && isComplete != true) {
+            return res.status(400).json({ error: 'Ingrese un valor boolean en isComplete' })
+        }
         connection.query("INSERT INTO tasks (title, description, isComplete) values (?,?,?)", [title, description, isComplete])
-        res.status(200).json({ message: "se ha agregado correctamente" })
+        res.status(201).json({ message: "se ha agregado correctamente" })
     } catch (error) {
         res.status(500).json({ error: 'Ha ocurrido un error' })
     }
@@ -41,22 +53,49 @@ const anadirTareas = async (req, res) => {
 
 // actualizar 
 const actualizarTareaId = async (req, res) => {
+    const id = req.params.id
+    if (!id) {
+        return res.status(400).json({ error: 'El Id es requerido para realizar una actualización' });
+    }
     try {
         const connection = await dataBase()
+        const [task] = await connection.query("SELECT * FROM tasks where id=?", [id])
+        if (task.length === 0) {
+            return res.status(404).json({ message: "el id solicitado no se encuentra en el sistema" })
+        }
         const { title, description, isComplete } = req.body
-        const id = req.params.id
+        if (title.trim() === '' || title.length > 225) {
+            return res.status(400).json({ error: 'El titulo no puede estar vacio y debe ser menor a 255 caracteres' });
+        }
+        if (description.trim() === '') {
+            return res.status(400).json({ error: 'La descripcion no debe estar vacia' });
+        }
+        if (isComplete != false && isComplete != true) {
+            return res.status(400).json({ error: 'Ingrese un valor boolean en isComplete' })
+        }
         const result = await connection.query("UPDATE tasks SET title = ? , description = ?, isComplete = ? WHERE id = ?", [title, description, isComplete, id])
-        res.send("se ha actualizado una tarea")
+        res.status(201).json({ message: "se ha actualizado correctamente" })
+
     } catch (error) {
         res.status(500).json({ error: 'Ha ocurrido un error' })
     }
 }
 
+
 //Eliminar tarea 
 const eliminarTareaporId = async (req, res) => {
+    const id = req.params.id
+    if (!id) {
+        return res.status(400).json({ error: 'El ID es requerido' });
+    }
+
     try {
         const connection = await dataBase()
-        const id = req.params.id
+        const [task] = await connection.query("SELECT * FROM tasks where id=?", [id])
+        if (task.length === 0) {
+            return res.status(404).json({ message: "el id solicitado no se encuentra en el sistema" })
+        }
+
         const results = await connection.query("DELETE FROM tasks WHERE id=?", id)
         res.send("se ha eliminado una tarea")
     } catch (error) {
